Item 7808: design review changes - layout changes to panels for data selections and plot options - remove run properties panel and put curve plot type in plot options panel
