+++ elisa/src/client/RunDetails/constants.ts
@@ -1,6 +1,5 @@
-
 export const X_AXIS_PROP = 'Concentration';
-export const Y_AXIS_PROP = 'Absorption';
+export const Y_AXIS_PROP = 'Absorption'; // TODO from Bob - Absorption could be changed to ECL
 export const SAMPLE_COL_NAME = 'SpecimenLsid/Property/SpecimenId';
 export const SAMPLE_COLUMN_NAMES = [SAMPLE_COL_NAME, 'WellLocation', Y_AXIS_PROP, X_AXIS_PROP, 'PlateName', 'Spot'];
-export const RUN_COLUMN_NAMES = ['Name','Created','CreatedBy/DisplayName','CurveFitMethod'];
+export const RUN_COLUMN_NAMES = ['Name','CurveFitMethod'];
